Stub for automatic role reveal

+++ config.py
@@ -74,15 +74,14 @@ class Config:
         
             self.mediavida_user = raw_config.loc["mediavida_user", "value"]
             self.mediavida_pwd  = raw_config.loc["mediavida_password", "value"]
-
             self.update_time    = int(raw_config.loc["update_time_seconds", "value"])
             self.posts_until_update = int(raw_config.loc["push_vote_count_interval", "value"])
             self.votes_until_update = int(raw_config.loc["votes_until_update", "value"])
             self.reveal_day_kill = bool(int(raw_config.loc["reveal_day_kill", "value"]))
             self.reveal_eod_lynch = bool(int(raw_config.loc["reveal_eod_lynch", "value"]))
             self.reveal_lynch = bool(int(raw_config.loc["reveal_lynch", "value"]))
-            self.day_duration = abs(int(raw_config.loc["day_hours", "value"]))
-            self.night_duration = abs(int(raw_config.loc["night_hours", "value"]))
+            self.day_duration = abs(float(raw_config.loc["day_hours", "value"]))
+            self.night_duration = abs(float(raw_config.loc["night_hours", "value"]))
 
             if self.update_time < 10:
                 self.update_time = 10
